Reestructuracion de repo
