+++ Controllers/CreacionController.cs
@@ -1,6 +1,6 @@
-﻿using backendAlquimia.Data.Entities;
+﻿using backendAlquimia.alquimia.Data;
 using backendAlquimia.Models;
-using backendAlquimia.Services.Interfaces;
+using backendAlquimia.alquimia.Services.Interfaces;
 using Microsoft.AspNetCore.Mvc;
 
 namespace backendAlquimia.Controllers
@@ -18,66 +18,66 @@ namespace backendAlquimia.Controllers
             _formulaService = formulaService;
         }
         // tiene que ser la primera solicitud.
-        [HttpGet("notasDeFondo")]
-        public async Task<ActionResult<IEnumerable<Nota>>> GetNotasFondo()
-        {
-            List<NotasPorFamiliaDTO> notas = await _notaService.ObtenerNotasDeFondoAgrupadasPorFamiliaAsync();
-            return Ok(notas);
-        }
+        //[HttpGet("notasDeFondo")]
+        //public async Task<ActionResult<IEnumerable<Nota>>> GetNotasFondo()
+        //{
+        //    List<NotasPorFamiliaDTO> notas = await _notaService.ObtenerNotasDeFondoAgrupadasPorFamiliaAsync();
+        //    return Ok(notas);
+        //}
 
-        [HttpGet("notasDeCorazon")]
-        public async Task<ActionResult<IEnumerable<Nota>>> GetNotasCorazon()
-        {
-            List<NotasPorFamiliaDTO> notas = await _notaService.ObtenerNotasDeCorazonAgrupadasPorFamiliaAsync();
-            return Ok(notas);
-        }
+        //[HttpGet("notasDeCorazon")]
+        //public async Task<ActionResult<IEnumerable<Nota>>> GetNotasCorazon()
+        //{
+        //    List<NotasPorFamiliaDTO> notas = await _notaService.ObtenerNotasDeCorazonAgrupadasPorFamiliaAsync();
+        //    return Ok(notas);
+        //}
 
-        [HttpGet("notasDeSalida")]
-        public async Task<ActionResult<IEnumerable<Nota>>> GetNotasSalida()
-        {
-            List<NotasPorFamiliaDTO> notas = await _notaService.ObtenerNotasDeSalidaAgrupadasPorFamiliaAsync();
-            return Ok(notas);
-        }
+        //[HttpGet("notasDeSalida")]
+        //public async Task<ActionResult<IEnumerable<Nota>>> GetNotasSalida()
+        //{
+        //    List<NotasPorFamiliaDTO> notas = await _notaService.ObtenerNotasDeSalidaAgrupadasPorFamiliaAsync();
+        //    return Ok(notas);
+        //}
 
-        [HttpGet("grado")]
-        public async Task<IActionResult> ObtenerGradoCompatibilidad(int notaAId, int notaBId)
-        {
-            var grado = await _notaService.CalcularCompatibilidadAsync(notaAId, notaBId);
-            return Ok(new { grado });
-        }
+        //[HttpGet("grado")]
+        //public async Task<IActionResult> ObtenerGradoCompatibilidad(int notaAId, int notaBId)
+        //{
+        //    var grado = await _notaService.CalcularCompatibilidadAsync(notaAId, notaBId);
+        //    return Ok(new { grado });
+        //}
 
-        [HttpPost("sugerencias")]
-        public async Task<IActionResult> ObtenerNotasCompatiblesAsync([FromBody] NotasSeleccionadasDTO dto)
-        {
-            var compatibles = await _notaService.ObtenerNotasCompatiblesAsync(dto.ListaDeIdsSeleccionadas, dto.Sector);
-            return Ok(compatibles);
-        }
+        //[HttpPost("sugerencias")]
+        //public async Task<IActionResult> ObtenerNotasCompatiblesAsync([FromBody] NotasSeleccionadasDTO dto)
+        //{
+        //    var compatibles = await _notaService.ObtenerNotasCompatiblesAsync(dto.ListaDeIdsSeleccionadas, dto.Sector);
+        //    return Ok(compatibles);
+        //}
 
-        [HttpGet("intensidad")]
-        public async Task<IActionResult> GetIntensidad()
-        {
-            List<IntensidadDTO> intensidad = await _formulaService.ObtenerIntensidadAsync();
-            return Ok(intensidad);
-        }
+        //[HttpGet("intensidad")]
+        //public async Task<IActionResult> GetIntensidad()
+        //{
+        //    List<IntensidadDTO> intensidad = await _formulaService.ObtenerIntensidadAsync();
+        //    return Ok(intensidad);
+        //}
 
-        [HttpPost("formular")]
-        public async Task<IActionResult> FinalizarCreacion([FromBody] POSTFormulaDTO dto)
-        {
-            GETFormulaDTO formulaGuardada = await _formulaService.guardar(dto);
-            return CreatedAtAction(
-                nameof(ObtenerFormulaPorId),
-                new { id = formulaGuardada.Id },
-                formulaGuardada
-                );
-        }
+        //[HttpPost("formular")]
+        //public async Task<IActionResult> FinalizarCreacion([FromBody] POSTFormulaDTO dto)
+        //{
+        //    GETFormulaDTO formulaGuardada = await _formulaService.guardar(dto);
+        //    return CreatedAtAction(
+        //        nameof(ObtenerFormulaPorId),
+        //        new { id = formulaGuardada.Id },
+        //        formulaGuardada
+        //        );
+        //}
 
-        [HttpGet("formulas/{id}")]
-        public async Task<IActionResult> ObtenerFormulaPorId(int id)
-        {
-            var formula = await _formulaService.ObtenerPorId(id);
-            if (formula == null) return NotFound();
-            return Ok(formula);
-        }
+        //[HttpGet("formulas/{id}")]
+        //public async Task<IActionResult> ObtenerFormulaPorId(int id)
+        //{
+        //    var formula = await _formulaService.ObtenerPorId(id);
+        //    if (formula == null) return NotFound();
+        //    return Ok(formula);
+        //}
