--- Data/Entities/Envase.cs
@@ -1,11 +1,0 @@
-﻿namespace backendAlquimia.Data.Entities
-{
-    public class Envase
-    {
-        public int Id { get; set; }
-        public string Forma { get; set; }
-        public int Volumen { get; set; }
-        public int IdEtiqueta { get; set; }
-        public Etiqueta Etiqueta { get; set; } = new();
-    }
-}
