--- Data/Entities/Esencia.cs
@@ -1,10 +1,0 @@
-﻿namespace backendAlquimia.Data.Entities
-{
-    public class Esencia : Producto
-    {
-        public int Id { get; set; }
-        public int IdNota { get; set; }
-
-
-    }
-}
