--- Data/Entities/Etiqueta.cs
@@ -1,12 +1,0 @@
-﻿namespace backendAlquimia.Data.Entities
-{
-    public class Etiqueta
-    {
-        public int Id { get; set; }
-        public string Texto { get; set; }
-        public string ColorTexto { get; set; }
-        public string ColorEtiqueta { get; set; }
-        public string Tipografia { get; set; }
-        public int AlturaLetra { get; set; }
-    }
-}
