--- Data/Entities/Intensidad.cs
@@ -1,9 +1,0 @@
-﻿using backendAlquimia.Data.Entities;
-
-public class Intensidad
-{
-    public int Id { get; set; }
-    public string Nombre { get; set; }
-    public List<Formula> Formulas { get; set; }
-
-}
