--- Data/Entities/Pedido.cs
@@ -1,8 +1,0 @@
-﻿namespace backendAlquimia.Data.Entities
-{
-    public class Pedido
-    {
-        public int Id { get; set; }
-        public List<Producto> Productos { get; set; } = new();
-    }
-}
