--- Data/Entities/Rol.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-
-namespace backendAlquimia.Data.Entities
-{
-    public class Rol : IdentityRole<int>
-    {
-
-    }
-}
