--- Data/Entities/TipoProducto.cs
@@ -1,10 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace backendAlquimia.Data.Entities;
-
-public class TipoProducto
-{
-   
-    public int Id { get; set; }
-    public string Description { get; set; }
-}
