+++ Seed/ProductoSeeder.cs
@@ -1,5 +1,5 @@
-﻿using backendAlquimia.Data;
-using backendAlquimia.Data.Entities;
+﻿using backendAlquimia.alquimia.Data;
+//using backendAlquimia.alquimia.Data.Entities;
 using Microsoft.EntityFrameworkCore;
 
 namespace backendAlquimia.Seed
@@ -9,20 +9,20 @@ namespace backendAlquimia.Seed
 
         public static async Task SeedTiposProductoAsync(IServiceProvider services)
         {
-            using var scope = services.CreateScope();
-            var context = scope.ServiceProvider.GetRequiredService<AlquimiaDbContext>();
+        //    using var scope = services.CreateScope();
+        //    var context = scope.ServiceProvider.GetRequiredService<AlquimiaDbContext>();
 
-            if (!await context.TiposProducto.AnyAsync())
-            {
-                await context.TiposProducto.AddRangeAsync(new List<TipoProducto>
-        {
-            new TipoProducto { Description = "Esencias puras" }, 
-            new TipoProducto { Description = "Alcohol" },
-            new TipoProducto { Description = "Frascos vidrio" }
-        });
+        //    if (!await context.TiposProducto.AnyAsync())
+        //    {
+        //        await context.TiposProducto.AddRangeAsync(new List<TipoProducto>
+        //{
+        //    new TipoProducto { Description = "Esencias puras" }, 
+        //    new TipoProducto { Description = "Alcohol" },
+        //    new TipoProducto { Description = "Frascos vidrio" }
+        //});
 
-                await context.SaveChangesAsync();
-            }
+        //        await context.SaveChangesAsync();
+            //}
         }
     }
 }
