+++ Seed/RoleSeeder.cs
@@ -1,4 +1,4 @@
-﻿using backendAlquimia.Data.Entities;
+﻿//using backendAlquimia.alquimia.Data.Entities;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.AspNetCore.Identity;
 
@@ -8,17 +8,17 @@ namespace backendAlquimia.Seed
     {
         private static readonly string[] Roles = { "Admin", "Creador", "Proveedor" };
 
-        public static async Task SeedRolesAsync(IServiceProvider serviceProvider)
-        {
-            var roleManager = serviceProvider.GetRequiredService<RoleManager<Rol>>();
+        //public static async Task SeedRolesAsync(IServiceProvider serviceProvider)
+        //{
+        //    var roleManager = serviceProvider.GetRequiredService<RoleManager<Rol>>();
 
-            foreach (var role in Roles)
-            {
-                if (!await roleManager.RoleExistsAsync(role))
-                {
-                    await roleManager.CreateAsync(new Rol { Name = role });
-                }
-            }
-        }
+        //    foreach (var role in Roles)
+        //    {
+        //        if (!await roleManager.RoleExistsAsync(role))
+        //        {
+        //            await roleManager.CreateAsync(new Rol { Name = role });
+        //        }
+        //    }
+        //}
     }
 }
