+++ Seed/UserSeeder.cs
@@ -1,4 +1,4 @@
-﻿using backendAlquimia.Data.Entities;
+﻿//using backendAlquimia.alquimia.Data.Entities;
 using Microsoft.AspNetCore.Identity;
 
 namespace backendAlquimia.Seed
@@ -7,26 +7,26 @@ namespace backendAlquimia.Seed
     {
         public static async Task SeedAdminAsync(IServiceProvider services)
         {
-            var userManager = services.GetRequiredService<UserManager<Usuario>>();
-            var email = "admin@alquimia.com";
-            var password = "Admin123!"; 
+            //var userManager = services.GetRequiredService<UserManager<Usuario>>();
+            //var email = "admin@alquimia.com";
+            //var password = "Admin123!"; 
 
-            var existingUser = await userManager.FindByEmailAsync(email);
-            if (existingUser == null)
-            {
-                var admin = new Usuario
-                {
-                    UserName = email,
-                    Email = email,
-                    Name = "Administrador"
-                };
+            //var existingUser = await userManager.FindByEmailAsync(email);
+            //if (existingUser == null)
+            //{
+            //    var admin = new Usuario
+            //    {
+            //        UserName = email,
+            //        Email = email,
+            //        Name = "Administrador"
+            //    };
 
-                var result = await userManager.CreateAsync(admin, password);
-                if (result.Succeeded)
-                {
-                    await userManager.AddToRoleAsync(admin, "Admin");
-                }
-            }
+            //    var result = await userManager.CreateAsync(admin, password);
+            //    if (result.Succeeded)
+            //    {
+            //        await userManager.AddToRoleAsync(admin, "Admin");
+            //    }
+            //}
         }
     }
 }
