--- Services/Interfaces/IJwtService.cs
@@ -1,9 +1,0 @@
-﻿using backendAlquimia.Data.Entities;
-
-namespace backendAlquimia.Services.Interfaces
-{
-    public interface IJwtService
-    {
-        string GenerateToken(Usuario user, IList<string> roles);
-    }
-}
