+++ alquimia.Services/Services/Interfaces/IFormulaService.cs
@@ -1,0 +1,11 @@
+﻿//using backendAlquimia.Models;
+
+namespace backendAlquimia.alquimia.Services.Interfaces
+{
+    public interface IFormulaService
+    {
+        //Task<GETFormulaDTO> guardar(POSTFormulaDTO dto);
+        //Task<List<IntensidadDTO>> ObtenerIntensidadAsync();
+        //Task<GETFormulaDTO> ObtenerPorId(int id);
+    }
+}
