+++ alquimia.Services/Services/Interfaces/IJwtService.cs
@@ -1,0 +1,9 @@
+﻿//using backendAlquimia.alquimia.Data.Entities;
+
+namespace backendAlquimia.alquimia.Services.Interfaces
+{
+    public interface IJwtService
+    {
+        //string GenerateToken(Usuario user, IList<string> roles);
+    }
+}
