+++ alquimia.Services/Services/Interfaces/INotaService.cs
@@ -1,0 +1,15 @@
+﻿//using backendAlquimia.Models;
+
+namespace backendAlquimia.alquimia.Services.Interfaces
+{
+    public interface INotaService
+    {
+        //Task<List<NotasPorFamiliaDTO>> ObtenerNotasDeSalidaAgrupadasPorFamiliaAsync();
+        //Task<List<NotasPorFamiliaDTO>> ObtenerNotasDeCorazonAgrupadasPorFamiliaAsync();
+        //Task<List<NotasPorFamiliaDTO>> ObtenerNotasDeFondoAgrupadasPorFamiliaAsync();
+        //Task<int> CalcularCompatibilidadAsync(int notaAId, int notaBId);
+        //Task<bool> EsCompatibleConSeleccionAsync(int nuevaNotaId, List<int> seleccionadasIds);
+        //Task<List<NotasPorFamiliaDTO>> ObtenerNotasCompatiblesAsync(List<int> seleccionadasIds, string Sector);
+
+    }
+}
