+++ backendAlquimia.sln
@@ -1,9 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35514.174 d17.12
+VisualStudioVersion = 17.12.35514.174
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "backendAlquimia", "backendAlquimia.csproj", "{FB5350E9-8300-4EC4-A958-135CAB6494E9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "alquimia.Services", "alquimia.Services\alquimia.Services.csproj", "{9EFCEEDD-699D-4682-989F-5CF0857F9808}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "alquimia.Data", "alquimia.Data\alquimia.Data.csproj", "{0428B043-1168-48FA-AEA9-D7ECD37EE761}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +19,14 @@ Global
 		{FB5350E9-8300-4EC4-A958-135CAB6494E9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FB5350E9-8300-4EC4-A958-135CAB6494E9}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FB5350E9-8300-4EC4-A958-135CAB6494E9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9EFCEEDD-699D-4682-989F-5CF0857F9808}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9EFCEEDD-699D-4682-989F-5CF0857F9808}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9EFCEEDD-699D-4682-989F-5CF0857F9808}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9EFCEEDD-699D-4682-989F-5CF0857F9808}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0428B043-1168-48FA-AEA9-D7ECD37EE761}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0428B043-1168-48FA-AEA9-D7ECD37EE761}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0428B043-1168-48FA-AEA9-D7ECD37EE761}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0428B043-1168-48FA-AEA9-D7ECD37EE761}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
